third_party/WebKit/LayoutTests/http/tests/serviceworker/chromium/navigation-preload-origin-trial-methods-expected.txt

CONSOLE MESSAGE: line 65: --normal SW--
disable() rejected with: AbortError: Failed to enable or disable navigation preload: The service worker script does not have a valid Navigation Preload Origin Trial token.
enable() rejected with: AbortError: Failed to enable or disable navigation preload: The service worker script does not have a valid Navigation Preload Origin Trial token.
getState() resolved with: {"enabled":false,"headerValue":"true"}
setHeaderValue() rejected with: AbortError: Failed to set navigation preload header: The service worker script does not have a valid Navigation Preload Origin Trial token.
log from SW
Activate event handler
registration.navigationPreload not found
Fetch event handler
event.preloadResponse not found

CONSOLE MESSAGE: line 76: --Origin-Trial enabled SW--
disable() rejected with: AbortError: Failed to enable or disable navigation preload: The Navigation Preload Origin Trial has ended.
enable() rejected with: AbortError: Failed to enable or disable navigation preload: The Navigation Preload Origin Trial has ended.
getState() resolved with: {"enabled":false,"headerValue":"true"}
setHeaderValue() rejected with: AbortError: Failed to set navigation preload header: The Navigation Preload Origin Trial has ended.
log from SW
Activate event handler
disable() rejected with: AbortError: Failed to enable or disable navigation preload: The Navigation Preload Origin Trial has ended.
enable() rejected with: AbortError: Failed to enable or disable navigation preload: The Navigation Preload Origin Trial has ended.
getState() resolved with: {"enabled":false,"headerValue":"true"}
setHeaderValue() rejected with: AbortError: Failed to set navigation preload header: The Navigation Preload Origin Trial has ended.
Fetch event handler
event.preloadResponse resolved with: undefined

This is a testharness.js-based test.
PASS Calling Navigation preload related methods for normal SW. 
PASS Calling Navigation preload related methods for Origin-Trial enabled SW. 
Harness: the test ran to completion.
